chore : Memory game 파일 refactoring
