+++ src/components/MemoryGame/MemoryGameBoard.tsx
@@ -1,8 +1,8 @@
 import { Box, Text, Grid } from '@chakra-ui/react';
 import { Dispatch, SetStateAction } from 'react';
-import { GRID_ITEM_COUNT, COLOR } from '../constants/memory';
+import { GRID_ITEM_COUNT, COLOR } from '../../constants/memory';
 import MemoryGameTimer from './MemoryGameTimer';
-import useScore from '../hooks/useScore';
+import useScore from '../../hooks/useScore';
 
 interface Props {
   stage: number;
