+++ src/components/MemoryGame/MemoryGameTimer.tsx
@@ -1,7 +1,7 @@
 import { Progress } from '@chakra-ui/react';
 import { useEffect, Dispatch, SetStateAction } from 'react';
-import useTimer, { TimerHookProps } from '../hooks/useTimer';
-import { GAME_STATE } from '../constants/memory';
+import useTimer, { TimerHookProps } from '../../hooks/useTimer';
+import { GAME_STATE } from '../../constants/memory';
 interface Props {
   setGameState: Dispatch<SetStateAction<string>>;
 }
